requirements.txt
# 电子会计档案管理系统依赖包
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
elasticsearch==8.9.0
redis==4.6.0
celery==5.3.2
Pillow==10.0.0
PyPDF2==3.0.1
reportlab==4.0.4
cryptography==41.0.4
hashlib2==1.0.1
bcrypt==4.0.1
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
jieba==0.42.1
requests==2.31.0
python-multipart==0.0.6
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
flasgger==0.9.7.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
jieba==0.42.1
hashlib2==1.0.1